cybernetics_core/requirements.txt
aiohttp
python-magic
structlog>=23.2.0
pydantic>=2.4.2
pydantic-settings>=2.1.0
langchain-ollama>=0.1.0           # New official Ollama integration
langchain-community>=0.2.1        # For updated imports
langchain-experimental>=0.0.49    # Required for SemanticChunker
prometheus-client>=0.17.1
qdrant-client>=1.8.3
redis>=4.6.0